init.target.rc: create netmgr recovery folder and file
To handle netmgr restart scenario and ensure call state
is cleaned up we need to cache some state on the device.
The cached information will be used to cleanup IPSec and
routing rules and bring back netmgr to a clean state to
allow subsequent iWLAN calls to work correctly.

The cached information will be saved in a file called
'recovery_info' which will be created via the init
script under a new folder which will also be created
by the init script.

CRs-Fixed: 2225506

Change-Id: I109ab059f9bf31e797498d702fd37b3e5d797b6b