init.target.rc: set vendor.usb.controller on charger
In charging mode, USB may be connected to a PC port. In order
to charge the battery at up to 500mA (or 900mA in the case of
SuperSpeed), the USB configuration needs to enumerate. For the
composition to apply, the correct controller device needs to be
specified similar to normal boot. Add vendor.usb.controller
property to 'on charger' trigger to allow the correct USB instance
to be started.

Change-Id: I819e6a7b324b9ed19e8283bc7024e1293796f1bf
Signed-off-by: gesangtome <gesangtome@foxmail.com>