init.target.rc: set sys.usb.controller on charger
In charging mode, USB may be connected to a PC port. In order
to charge the battery at up to 500mA (or 900mA in the case of
SuperSpeed), the USB configuration needs to enumerate. For the
composition to apply, the correct controller device needs to be
specified similar to normal boot. Add sys.usb.controller property
to 'on charger' trigger to allow the correct USB instance
to be started.

Change-Id: I70c092c3a0c4223a838f3f4ae7fccb622e8e474b
Signed-off-by: gesangtome <gesangtome@foxmail.com>